back to vscode - windsurf's autocomplete is just in the way.
sometimes I just want to write code and not have AI do anything. VSCode
with Cline+Supermaven is a better combo for me.

But this also makes me want to move to neovim fulltime because I am just
configuring vscode to be like neovim. I need to map my keybindings to
match.